requirements.txt
flask
flask_cors
PyMuPDF
sentence_transformers
google-generativeai
numpy
chromadb